miniups-backend/mini-ups/requirements.txt
SQLAlchemy
psycopg2-binary

protobuf==3.20.3